+++ models/energy_supply/energy_supply.py
@@ -197,22 +197,16 @@ class EnergySupplyWrapper(SectorModel):
         write_input_timestep(gasload, "gasload",
                              now, region_names, interval_names)
 
-
-    def get_model_executable(self):
-        """Return path of current python interpreter
-        """
-        nismod_dir = os.path.join(os.path.dirname(__file__), '..', '..')
-        return os.path.join(nismod_dir, 'energy_supply', 'model', 'Energy_Supply_Master.mos')
-
     def run_the_model(self):
         """Run the model
         """
         nismod_dir = os.path.join(os.path.dirname(__file__), '..', '..')
-        os.environ["ES_PATH"] = str(os.path.abspath(os.path.join(
-            nismod_dir, 'energy_supply', 'model')))
+        model_dir = os.path.join(nismod_dir, 'install', 'energy_supply')
+
+        os.environ["ES_PATH"] = str(os.path.abspath(model_dir))
         self.logger.debug("\n\n***Running the Energy Supply Model***\n\n")
-        arguments = self.get_model_executable()
-        self.logger.debug(check_output(['mosel', 'exec', arguments]))
+        model_path = os.path.join(model_dir, 'Energy_Supply_Master.exe')
+        self.logger.debug(check_output([model_path]))
 
     def retrieve_outputs(self, data, now):
         """Retrieves results from the model
@@ -1097,3 +1091,43 @@ def write_input_timestep(input_data, parameter_name, year,
     # Close communication with the database
     cur.close()
     conn.close()
+
+
+class EnergySupplyDevWrapper(EnergySupplyWrapper):
+    """Monkey patches run model method to call development version of the model
+
+    """
+    def run_the_model(self):
+        """Run the model
+        """
+        nismod_dir = os.path.join(os.path.dirname(__file__), '..', '..')
+        model_dir = os.path.join(nismod_dir, 'energy_supply', 'model')
+
+        os.environ["ES_PATH"] = str(os.path.abspath(model_dir))
+        self.logger.debug("\n\n***Running the Energy Supply Model***\n\n")
+        model_path = os.path.join(nismod_dir, 'energy_supply', 'model', 'Energy_Supply_Master.mos')
+        self.logger.debug(check_output(['mosel', 'exec', model_path]))
+
+
+class EnergySupplyToyWrapper(EnergySupplyWrapper):
+    """Monkey patches methods in full version of wrapper with minimal dimension
+    definitions
+    """
+
+    def get_names(self, name, spatial_name='energy_hub_min', temporal_name='seasonal_week'):
+        """Get region and interval names for a given input
+        """
+        return super().get_names(name, spatial_name, temporal_name)
+
+    def get_gasload(self, data, now):
+        gasload = data.get_data('gasload')
+        region_names, interval_names = self.get_names("gasload", spatial_name='gas_nodes_minimal')
+        self.logger.debug('Writing %s to database', "gasload")
+        write_input_timestep(gasload, "gasload", now, region_names, interval_names)
+
+
+class EnergySupplyDevToyWrapper(EnergySupplyDevWrapper, EnergySupplyToyWrapper):
+    """Monkey patches run model method to call development version of the minimal 
+    model
+    
+    """
